game-engine-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.1
requests==2.32.3
openai==1.51.2
bcrypt==4.2.0
PyJWT==2.9.0


